editor.php update started
