+++ PHP/functionEditor.php
@@ -1,0 +1,28 @@
+<?php  
+session_start();
+include("connection.php");
+$user_id = $_SESSION["user_id"];
+
+
+if(isset($_POST["action"])){
+    if($_POST["action"] == "about"){
+        // echo "user id : " . $user_id . "..";
+        updateAbout($user_id);
+    }
+}
+
+function updateAbout($user_id){
+    global $connection;
+
+    $aboutUser = $_POST["about_user"];
+    $yearsOfExperience = $_POST["years_of_experience"];
+    $completedProjects = $_POST["completed_projects"];
+    $companiesWorked = $_POST["companies_worked"];
+
+    $stmt = $connection->prepare("UPDATE about_tab_tb SET aT_about_user = ?, aT_Yo_Exp = ?, aT_No_Projects = ?, aT_No_companies = ? WHERE aT_user_id = ?");
+    $stmt->bind_param("ssssi", $aboutUser, $yearsOfExperience, $completedProjects, $companiesWorked, $user_id);
+    $stmt->execute();
+
+    echo "Saved Successfully";
+
+}
